+++ app/src/main/java/io/casehub/devtown/app/PrReviewResource.java
@@ -1,5 +1,8 @@
 package io.casehub.devtown.app;
 
+import io.casehub.devtown.review.PrPayload;
+import io.casehub.devtown.review.PrReviewApplicationService;
+import io.casehub.devtown.review.PrReviewOutcome;
 import jakarta.inject.Inject;
 import jakarta.ws.rs.Consumes;
 import jakarta.ws.rs.POST;
