+++ review/src/main/java/io/casehub/devtown/review/PrPayload.java
@@ -1,3 +1,3 @@
-package io.casehub.devtown.app;
+package io.casehub.devtown.review;
 
 public record PrPayload(String repo, int prNumber, String headSha, int linesChanged) {}
