+++ review/src/main/java/io/casehub/devtown/review/PrReviewApplicationService.java
@@ -1,4 +1,4 @@
-package io.casehub.devtown.app;
+package io.casehub.devtown.review;
 
 public interface PrReviewApplicationService {
     PrReviewOutcome review(PrPayload pr);
