+++ review/src/main/java/io/casehub/devtown/review/PrReviewOutcome.java
@@ -1,4 +1,4 @@
-package io.casehub.devtown.app;
+package io.casehub.devtown.review;
 
 import java.util.List;
